+++ biopipen/scripts/scrna/SeuratPreparing-doublet_detection.R
@@ -158,6 +158,7 @@ plot_dd <- function(sobj, dd, detector) {
         sobj, group.by = paste0(detector, "_DropletType"), order = "doublet",
         cols = c("#333333", "#FF3333"), pt.size = 0.8, alpha = 0.5)
     ggsave(dimp, filename = file.path(plotsdir, paste0(detector, "_dimplot.png")))
+    ggsave(dimp, filename = file.path(plotsdir, paste0(detector, "_dimplot.pdf")))
 }
 
 filter_dd <- function(sobj, dd, detector) {
@@ -184,15 +185,17 @@ report_dd <- function(detector) {
 
     if (detector == "DoubletFinder") {
         add_report(
-            list(name = "pK vs BCmetric", src = file.path(plotsdir, "pK_BCmetric.png")),
-            list(name = "Dimension Reduction Plot", src = file.path(plotsdir, "DoubletFinder_dimplot.png")),
+            list(name = "pK vs BCmetric", src = file.path(plotsdir, "DoubletFinder_pK_BCmetric.png")),
+            list(name = "Dimension Reduction Plot", src = file.path(plotsdir, "DoubletFinder_dimplot.png"),
+                 download = file.path(plotsdir, "DoubletFinder_dimplot.pdf")),
             ui = "table_of_images",
             h1 = "DoubletFinder Results",
             h2 = "Plots"
         )
     } else {
         add_report(
-            list(name = "Dimension Reduction Plot",src = file.path(plotsdir, "scDblFinder_dimplot.png")),
+            list(name = "Dimension Reduction Plot", src = file.path(plotsdir, "scDblFinder_dimplot.png"),
+                 download = file.path(plotsdir, "scDblFinder_dimplot.pdf")),
             ui = "table_of_images",
             h1 = "scDblFinder Results",
             h2 = "Plots"