+++ tests/test_scrna/ScFGSEA/test.py
@@ -63,7 +63,11 @@ class ScFGSEAEach(ScFGSEA):
 
 
 def pipeline():
-    return get_pipeline(__file__).set_start(PrepareSeurat)
+    return (
+        get_pipeline(__file__)
+        # get_pipeline(__file__, enable_report=True)
+        .set_start(PrepareSeurat)
+    )
 
 
 def testing(pipen):